+++ packages/backend/package.json
@@ -1,6 +1,7 @@
 {
   "name": "backend",
   "version": "1.0.0",
+  "type":"module",
   "main": "backend.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
